// frontend/contracts/observer/v1/models/ContestEventType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum ContestEventType {
    CONTEST_ANNOUNCEMENT_CREATED = 'contest.announcement_created',
    CONTEST_ANNOUNCEMENT_DELETED = 'contest.announcement_deleted',
    CONTEST_CLARIFICATION_CREATED = 'contest.clarification_created',
    CONTEST_CLARIFICATION_ANSWERED = 'contest.clarification_answered',
}