// frontend/contracts/observer/v1/models/MessageContestAnnouncementCreated.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MessageContestAnnouncementCreated = {
    id: string;
    contest_id: string;
    problem_id?: string;
    problem_title?: string;
    problem_letter?: string;
    author_id?: string;
    author_username: string;
    title: string;
    body: string;
    created_at: string;
};