// frontend/contracts/observer/v1/models/MessageContestClarificationCreated.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MessageContestClarificationCreated = {
    id: string;
    contest_id: string;
    problem_id?: string;
    problem_title?: string;
    problem_letter?: string;
    user_id: string;
    username: string;
    question: string;
    created_at: string;
};