// frontend/contracts/observer/v1/models/MessageContestClarificationAnswered.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MessageContestClarificationAnswered = {
    id: string;
    contest_id: string;
    problem_id?: string;
    problem_title?: string;
    problem_letter?: string;
    user_id: string;
    username: string;
    question: string;
    answer: string;
    answered_by?: string;
    answered_by_username?: string;
    answered_at: string;
};